+++ airbyte-integrations/connectors/source-postgres/src/test/java/io/airbyte/integrations/source/postgres/CdcPostgresSourceTest.java
@@ -7,7 +7,11 @@ package io.airbyte.integrations.source.postgres;
 import static io.airbyte.integrations.debezium.internals.DebeziumEventUtils.CDC_DELETED_AT;
 import static io.airbyte.integrations.debezium.internals.DebeziumEventUtils.CDC_LSN;
 import static io.airbyte.integrations.debezium.internals.DebeziumEventUtils.CDC_UPDATED_AT;
-import static org.junit.jupiter.api.Assertions.*;
+import static org.junit.jupiter.api.Assertions.assertEquals;
+import static org.junit.jupiter.api.Assertions.assertFalse;
+import static org.junit.jupiter.api.Assertions.assertNotNull;
+import static org.junit.jupiter.api.Assertions.assertNull;
+import static org.junit.jupiter.api.Assertions.assertTrue;
 
 import com.fasterxml.jackson.databind.JsonNode;
 import com.fasterxml.jackson.databind.node.ObjectNode;
@@ -199,7 +203,7 @@ public class CdcPostgresSourceTest extends CdcSourceTest {
   @Test
   void testCheckWithoutReplicationPermission() throws Exception {
     createCleanUser();
-    JsonNode test_config = getConfig(dbName, cleanUserName, cleanUserPassword);
+    final JsonNode test_config = getConfig(dbName, cleanUserName, cleanUserPassword);
     final AirbyteConnectionStatus status = source.check(test_config);
     assertEquals(AirbyteConnectionStatus.Status.FAILED, status.getStatus());
     assertEquals(String.format(ConnectorExceptionUtil.COMMON_EXCEPTION_MESSAGE_TEMPLATE,
@@ -310,6 +314,34 @@ public class CdcPostgresSourceTest extends CdcSourceTest {
   @Override
   protected String randomTableSchema() {
     return MODELS_SCHEMA + "_random";
+  }
+
+  @Test
+  void testDiscoverFiltersNonPublication() throws Exception {
+    // Drop the default publication (which is created for all tables). Create a publication for the models table. By default,
+    // the tests create a models_schema.models table and models_schema_random.models_random table. We will create a publication
+    // for one of the tests and assert that both streams end up in the catalog. However, the stream that is not associated with
+    // a publication should only have SyncMode.FULL_REFRESH as a supported sync mode.
+    database.query(ctx -> ctx.execute("DROP PUBLICATION " + PUBLICATION + ";"));
+    database.query(ctx ->
+        ctx.execute(String.format("CREATE PUBLICATION " + PUBLICATION + " FOR TABLE %s.%s", MODELS_SCHEMA, "models")));
+
+    final AirbyteCatalog catalog = source.discover(config);
+    assertEquals(catalog.getStreams().size(), 2);
+    final AirbyteStream streamInPublication =
+        catalog.getStreams().stream().filter(stream -> stream.getName().equals("models")).findFirst().get();
+    final AirbyteStream streamNotInPublication =
+        catalog.getStreams().stream().filter(stream -> !stream.getName().equals("models")).findFirst().get();
+
+    // The stream that has an associated publication should have support for source-defined incremental sync.
+    assertEquals(streamInPublication.getSupportedSyncModes(), List.of(SyncMode.FULL_REFRESH, SyncMode.INCREMENTAL));
+    assertFalse(streamInPublication.getSourceDefinedPrimaryKey().isEmpty());
+    assertTrue(streamInPublication.getSourceDefinedCursor());
+
+    // The stream that does not have an associated publication should not have support for source-defined incremental sync.
+    assertEquals(streamNotInPublication.getSupportedSyncModes(), List.of(SyncMode.FULL_REFRESH));
+    assertTrue(streamNotInPublication.getSourceDefinedPrimaryKey().isEmpty());
+    assertFalse(streamNotInPublication.getSourceDefinedCursor());
   }
 
   @Test